+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Changes
 
 * [#663](https://github.com/clojure-emacs/cider-nrepl/pull/663): Fix non-recognized `recur` inside `case` in the debugger.
+* [#664](https://github.com/clojure-emacs/cider-nrepl/pull/664): Fix continue-all in conditional break.
 
 ## 0.23.0 (2019-01-18)